server/pyproject.toml
[project]
name = "desarrollo"
version = "0.1.0"
description = "Trabajo Práctico Final de Desarrollo de Software"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "markdown>=3.7",
    "djangorestframework>=3.15.2",
    "coverage>=7.6.1",
    "pygments>=2.18.0",
    "python-dotenv>=1.0.1",
    "python-json-logger>=2.0.7",
    "requests>=2.32.3",
    "django-cors-headers>=4.5.0",
    "pillow>=11.0.0",
    "gunicorn>=23.0.0",
    "docutils>=0.21.2",
    "whitenoise>=6.8.2",
    "django-filter>=24.3",
    "django-ratelimit>=4.1.0",
    "colorlog>=6.9.0",
    "qrcode[pil]>=8.0",
    "ulid-py>=1.1.0",
    "datetime>=5.5",
    "psycopg[pool]>=3.2.3",
    "django-background-tasks>=1.2.8",
    "ruff>=0.7.4",
    "django-storages[google]>=1.14.4",
    "google-cloud-storage>=2.18.2",
    "dj-database-url>=2.3.0",
    "pathlib>=1.0.1",
    "python-decouple>=3.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
